+++ src/modules/validations/decorators/numbers/is-number.ts
@@ -4,7 +4,7 @@ import type {
   ValidationDecoratorDefinition,
 } from 'typings/validations.ts'
 
-import { defineValidationDecorator } from 'modules/validations/base/definitions/decorators.ts'
+import { defineCatalogValidationDecorator } from 'modules/validations/base/definitions/decorators.ts'
 import { defaultTransform } from './defaults.ts'
 
 /**
@@ -68,7 +68,7 @@ export const IsNumber: ValidationDecorator<
     validation = isNumber
   }
 
-  return defineValidationDecorator(validation, {
+  return defineCatalogValidationDecorator(validation, {
     transform: defaultTransform(transform),
     message: defaultMessage,
     ...opts,