+++ src/modules/validations/decorators/strings/is-number-string.ts
@@ -1,6 +1,6 @@
 import type { ValidationDecorator, ValidationDecoratorDefinition } from 'typings/validations.ts'
 
-import { defineValidationDecorator } from 'modules/validations/base/definitions/decorators.ts'
+import { defineCatalogValidationDecorator } from 'modules/validations/base/definitions/decorators.ts'
 import { match, matchArray } from './match.ts'
 import { NUMERIC_REGEX } from 'utils/regex.ts'
 
@@ -51,8 +51,8 @@ export const IsNumberString: ValidationDecorator = function (
     validation = isNumberString
   }
 
-  return defineValidationDecorator(validation, {
+  return defineCatalogValidationDecorator(validation, {
     message: defaultMessage,
     ...options,
-  })
+  }, { decorator: 'IsNumberString' })
 }